[Bayan] Added configs to separate package
