+++ AuthService/src/main/java/CS4337/Project/Config/AppConfig.java
@@ -1,0 +1,14 @@
+package CS4337.Project.Config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class AppConfig {
+
+    @Bean
+    public RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
+}
