+++ AuthService/src/main/java/CS4337/Project/Config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package CS4337.Project;
+package CS4337.Project.Config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
